fix build on arm
We have a 64 bit integer in one struct definition, and we're casting
another struct that contains nothing that large to it. The struct
that only contains smaller types only has a 32-bit alignment
constraint, but is being cast to something with a 64-bit alignment
constraint, so gcc complains.

Break the 64-bit item into high/low components and recombine/break
apart as required.