eos-paygd: Don't enforce PAYG if Secure Boot is off
If the current boot occurred with Secure Boot disabled, it doesn't make
sense to enforce PAYG because (a) it's easily circumventable by
replacing the initramfs and (b) Secure Boot has likely been disabled
because the machine is in one of two states:
1. It has been fully paid off and unlocked (though note that it's also
possible to have Secure Boot enabled on an unlocked machine, in which
case EOSPAYG_active will not be set).
2. It is being debugged by a developer who has used PKpriv to disable
Secure Boot.

This commit also adds a command line flag --skip-sb-check, so a
developer can have eos-paygd in enforcing mode even with Secure Boot
off (though adding the flag would require editing the initramfs or using
a specially created debug image).

https://phabricator.endlessm.com/T27735