eos-paygd: Only ping watchdog every minute
Currently we ping the watchdog timer every 10 seconds, but this is a bit
excessive given that it's a 20 minute timer. So increase the time
between pings to 60 seconds, and update the documentation to match the
changes made here[1]. I'm not sure exactly what the performance cost is
of this ping, which should just be in memory. We do know the timer stops
when the computer is suspended, so it's okay that pings won't happen
in that state.

[1] https://github.com/endlessm/linux/pull/502