Shut down 10 minutes after credit expiration
Unlike eos-paygd, gnome-shell is outside the perimeter of trust; it can
be replaced with a modified version by the user. And so it may not
enforce PAYG. Therefore it's necessary for eos-paygd to shut down the
computer after credit has expired and a grace period has elapsed, to
ensure the computer cannot be used in an expired state for very long. So
this commit changes EpgManagerService so that it starts a 10 minute
timer when the provider emits the Expired signal. If the timer reaches
zero, the computer is shut down. If the expiration time is extended
during those 10 minutes, the timer is cancelled. We also emit a D-Bus
signal called ImpendingShutdown, which will be used by gnome-shell to
inform the user about the shutdown timer, though the user-facing part
in gnome-shell is not being implemented yet.

The existing supported EpgProvider implementation (for Angaza) already
emits the expired signal on startup when no credit exists. Update the
docs for the expired signal to make it clear that future implementations
should do so as well; otherwise one need only reboot to get around this
protection.

https://phabricator.endlessm.com/T27133