eos-paygd: Activate LSM if enforcing PAYG
This commit makes eos-paygd activate the custom "endlesspayg" LSM in the
kernel, which provides various protections for the process such as
preventing it from being killed even by root. We activate it before the
root pivot while we're still in a trusted environment, and we only
activate if in enforcing mode. We're in enforcing mode if the current
boot is secure and the EFI variable EOSPAYG_active is set. If we're not
in enforcing mode either the machine has been paid off, or it is being
debugged, or it hasn't been provisioned yet.

https://phabricator.endlessm.com/T27050