eos-paygd: Only use the watchdog if enforcing
Now that we're only activating the LSM if in PAYG enforcing mode, I
think it makes sense to do the same for the watchdog device and only
open it if in enforcing mode. This means that when a PAYG machine has
been unlocked e.g. for debugging, one can attach gdb to eos-paygd and
walk away for 20 minutes without the computer forcibly powering off. The
reason we previously were activating the watchdog unconditionally was to
prevent any other process from using it, but that seems like an unlikely
risk that some process would open it, not ping it, and cause a poweroff
either accidentally or maliciously. There are plenty of other ways a
root user could cause the machine to power off.

https://phabricator.endlessm.com/T27050