Implement support for EOSPAYG_debug EFI variable
This EFI variable should prove useful when debugging PAYG systems which
in production are so locked down as to be difficult to debug. You can,
for example, set a bit to tell eos-paygd not to use the LSM, which means
the process will show up in /proc and can be attached to with gdb. This
variable cannot easily be set on a production system because the LSM
ensures that only eos-paygd can access EFI variables which have the
prefix "EOSPAYG". The expected flow is that the developer will build a
bootable EFI blob which unlocks the computer in question if the account
number on the computer matches the one compiled into the blob. See
https://phabricator.endlessm.com/T28290. Then, even if the primary OS
isn't booting, you could disable Secure Boot, boot from a USB drive, and
set EOSPAYG_debug using the instructions here:
https://phabricator.endlessm.com/w/software/pay-as-you-go/

https://phabricator.endlessm.com/T27734