eos-paygd: Properly retry to connect to D-Bus
Currently a computer running a payg image which does not have state data
and does not have EOSPAYG_active set abruptly shuts down after 20
minutes. This is not correct; it should behave as a non-PAYG computer.
The reason for the shutdown is that we don't use the watchdog device if
we're not enforcing PAYG, but we still have a call to
"g_main_context_iteration (NULL, TRUE)" in the retry loop used for
connecting to the system bus. Since there are no events to process
except the payg_sync_and_poweroff() timeout, g_main_context_iteration()
blocks for 20 minutes and the computer shuts down.

This commit changes the call to "g_main_context_iteration (NULL, FALSE)"
which means that it won't block when there are no events to process.
This means that when the watchdog is not used, the call will be a no-op,
unless 20 minutes pass without a successful D-Bus connection. And when
the watchdog is in use, it will be pinged only when 60 seconds have
passed, without blocking a D-Bus connection retry which happens every
second.

https://phabricator.endlessm.com/T28417