Trigger shutdown timer if expired on startup
Currently, we start a timer to shut down in 10 minutes if PAYG credit
runs out while the computer is running (to ensure we don't have to rely
on gnome-shell's enforcement for security). We should also start a
shutdown timer on startup, but this isn't working. This is because when
the EpgProvider object emits its "expired" signal on startup, it has not
yet been added to an EpgManagerService object. So when EpgManagerService
connects a callback to the expired signal, it only gets executed if
credit expires after that point.

The fix is to check if the provider is expired when it's being added as
a property to the EpgManagerService, which happens during construction.

https://phabricator.endlessm.com/T27133