Log a message before shutting down
Use g_message() to log a message before starting a 10 minute shutdown
timer (when credit runs out) and log another if the timer is cancelled.
If we used g_debug() or g_info() it would only show up in the log if
G_MESSAGES_DEBUG is set in the environment. I think this is worth always
logging.