eos-paygd: Allow writes to /dev/mmcblk*boot0
We previously made this change for eos-paygd-2.service (which runs under
PAYG >= 4) but did not apply it for PAYG 2 systems.

As a result, all PAYG writes would fail so a user who applied a code
would lose the credit upon reboot and, even worse from a security
perspective, be able to re-apply any codes not applied before they
upgraded to a version of EOS that included this bug. Thus, a user could
buy a long code and conveniently just reboot to re-apply it an unlimited
number of times.

https://phabricator.endlessm.com/T29201