efi: Return 0 from read if a variable exists but has 0 size
efivarfs will sometimes think a variable has been created when
it hasn't, and this will result in a variable having the
impossible size of 0 bytes in the fs.

Previously we returned a size of -1 for this. It makes more sense
to reserve the -1 case for "file not found" situations, so calling
code doesn't think a 0 byte variable is not present and can be
written without deleting the old one first.

https://phabricator.endlessm.com/T31137