readme: Make a note regarding shutdown during nv write
We've recently discussed the possibility of system shutdown interrupting
a write to non volatile storage. The conclusion is that eos-paygd is
safe against these interruptions, so we don't have to deal with
shutting the daemon down cleanly and all that entails.

However, the reason it's safe is because all currently in use nv storage
backends can handle interrupted writes safely (this does not include the
file backend, which is not used in the real world).

Make a note of this in the README so future developers realize both that
shutdown is safe, and that atomicity is a critical design criteria for
the storage backends to maintain this safety.

https://phabricator.endlessm.com/T27718