hwclock: Print the for failing to opening the RTC
This will aid in diagnosing when the RTC hardware is busted the daemon
enforces the PAYG lock and forces a shutdown. We have seen this failing
on some ECS EF20EA machines in the field, but without having access to
the machines we can't determine why it fails. We suspect a hardware
failure, but having the actual error printed on the logs will be help.

https://phabricator.endlessm.com/T32703