hwclock: Keep the RTC in UTC
Since when payg is enforced we disable the kernel SYSTOHC functionality
and prevent any other userspace process from accessing the RTC,
eos-paygd effectively becomes responsive for keeping the RTC up-to-date.

This code previously assumed the RTC was configured to use the local
time, which is not actually the case -- the RTC will always use UTC on
standalone Endless OS installations.

This commit changes eos-paygd behavior to always store the UTC time in
the RTC, and treat values read from the RTC as UTC time.

https://phabricator.endlessm.com/T32733