paygd: Don't exit immediately on SIGTERM / SIGINT
At this point the signal has been handled by the service object and its
mainloop has exited, but we need to do some error handling before
terminating the daemon.

Also, there is no need to log here that the daemon was terminated by a
signal, the signal handler installed by the service object already logs
which signal was received at the debug level.

https://phabricator.endlessm.com/T33088
https://phabricator.endlessm.com/T32815