paygd: Don't force poweroff if already in progress
If we request a poweroff via systemctl while a system shutdown is
already in progress, the systemctl command will exit with an error. In
this case we should not force a poweroff using the kernel API and just
let systemd finish the shutdown cleanly.

Unfortunatelly there is no easy and secure way to differentiate between
a shutdown already being in progress and a failure to request a
shutdown, so let's remove the fallback path with a hard shutdown to
avoid forcing it when we should not.  This should not cause any problems
since the kernel watchdog will still make sure a shutdown occurs
eventually.

https://phabricator.endlessm.com/T33088
https://phabricator.endlessm.com/T32815