paygd: Request poweroff if the service exits with any error
Handle all cases where the service object exits with an error by
requesting a poweroff after 20 minutes. The causes we expect to trigger
this logic are the provider data being erased or damaged, or paygd
losing the bus name (which can happen if the D-Bus server gets
terminated or restarted).

https://phabricator.endlessm.com/T33037