manager: Use the device ID as the PAYG account ID
When using our simplistic default backend the account ID is left blank.
We do, however, generate a device ID during provisioning, from the first
8 digits of the machine ID at provisioning time, which is used by the
vendor to identify the account associated with the device.

Let's save this ID to the disk during provisioning so paygd can load it
and display it in the UI as the account ID.

https://phabricator.endlessm.com/T33322