provision-phase-1: Remove tarballs from repository
Downloading tarballs from GitHub is not convenient for partners, and
having them as part of the repository is impractical for development
(need to update tarballs on every commit that touch to keep test suite
green).

Now that we are publishing them via a Jenkins job we can remove these
tarballs and the respective tests from the repository.

https://phabricator.endlessm.com/T33373