paygd: Write logs to a separate file
In systems with fragile storage (ex., SSD or eMMC) the journal is only
kept in RAM as does not persist over reboot. In such systems we still
want logs from eos-paygd to persist, without having to write all system
logs to persistent storage.

The solution for that is to create a separate log file for eos-paygd, in
addition to forwarding logs to the journal.

https://phabricator.endlessm.com/T33220