paygd: Add date stamp to log file name
Add the current date to the log file name, so the logs get split into
separate files when paygd is started on different days.

With this commit the log file is open on startup only, so the same file
may be used accross several days if there system / daemon is not
restarted.

https://phabricator.endlessm.com/T33220