paygd: Always open the log file at write time
Always re-construct the log file name and open it right before writing,
and close it immediately after using it. Since the log file name
contains a date stamp, this will split logs in per-day files.

This way we can easily rotate logs with a tmpfiles.d entry like the
following:

 # Files modified over 90 days ago are subject to time-based cleanup
 d /var/log/eos-payg 0755 root root m:90d

https://phabricator.endlessm.com/T33220