paygd: Ship tmpfile entry to maintain /var/log/eos-payg
This will instruct systemd to remove any files that have not been
created, modified or accessed in the last 90 days. Ideally we would have
used 'm:90d' so just opening a file for reading would not interfere with
its cleanup, but this is only supported on systemd 249+ and EOS 4.0.x
ships systemd 247.

The 'd' entry we are using also tells systemd to create the directory if
it does not exist. This will be the case for new installations, where
the system is initially unprovisioned so eos-paygd will not be started,
but for existing installations that are already provisioned eos-paygd
will most likely create it first.

https://phabricator.endlessm.com/T33220