eos-paygd: Fix file encoding of log file path
`g_strconcat()` is not encoding-aware, whereas the path passed in to
`g_io_channel_new_file()` should be in the file system encoding. Use
`g_build_filename()` to guarantee that’s the case, so that this will
work on the (very few) systems which don’t have UTF-8 as their file
system encoding.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>