service: Pass current time to provider on clock jumps
When a clock jump is detected, we already need to get the current
wallclock time to calculate the jump size. Let's pass this information
to the provider, so it can better decide how to handle the jump.

Also update the "manager" provider to use the time passed by caller in
its wallclock_time_changed implementation instead of calculating again.

https://phabricator.endlessm.com/T33873