manager: Fix the time to set_expiry_time on clock jumps
On commit ae1af69bbea1 "service: Pass current time to provider on clock
jumps" a parameter with the time of the jump was added to the
wallclock_time_changed family of functions. Its value is relative to
CLOCK_REALTIME, and in epg_manager_wallclock_time_changed we were
passing it directly to set_expiry_time, which expects a time value
relative to CLOCK_BOOTTIME. This commit fixes that issue.

Fixes: ae1af69bbea1 ("service: Pass current time to provider on clock jumps")

https://phabricator.endlessm.com/T34550