manager: Fix credit consumption on clock jumps
Previously we were setting the expiry time (thus the amount of credit)
to X seconds from now, where X was the size of the forward jump in
seconds (delta). This could add or reduce credit, depending on the size
of the jump.

This commit fixes the logic so we reduce the amount of credit by the
same amount that the clock moved forward.

https://phabricator.endlessm.com/T34550