manager: set_expire_time: Always use the passed base value
In set_expire_time the now_secs parameter is used for the caller to pass
the current time, to be used as a base to calculate the new expire time
by adding the requested amount of span_secs. Except that if the current
expiry time is in the future (the computer is unlocked), it instead uses
that expiry time as the base for the calculation, extending the amount
of credit by span_secs.

This state-dependent behaviour makes this function difficult to use, so
this commit changes it so it always used the value passed in now_secs as
the base for calculating the new expiry time, and moves logic that
checks whether the computer is unlocked or not to the only caller that
relies on it, extend_expiry_time.

This commit brings no functional changes.

https://phabricator.endlessm.com/T34550