Support 31-day codes
Paying on the same calendar day each month is predictable and standard
for other subscription models. But previously, doing this would either
mean issuing two codes in months with 31 days (one for 30 days, another
for 1 day) or the user running out of credit for a day.

By supporting 31-day codes, it is possible to either issue a 30-day or
31-day code depending on the month, or (more likely) to simply issue a
31-day code every month, leaving the user with 7-8 days' extra credit at
the end of each year.

(We could in principle add 28-day and 29-day codes to support February
and lunar calendars "natively", but there are only 5 more bits available
in this bitfield.)

https://phabricator.endlessm.com/T33556