manager: Adjust credit on systems with way too much
Due to a software bug, some systems ended up with aproximately 50 years
of credit. To recover from this situation, this commit adjusts systems
with more than 5 years but less than 100 years of credit to have 31 days
of credit.

These values were chosen because we do not expect a system to have 5+
years of credit by valid means, except if it is fully unlocked, when we
add G_MAXUINT64 seconds of credit, which is around 584 billion years.
And we reset to 31 days because that is the standard payment period for
deployments using the manager (aka phase1) provider.

https://phabricator.endlessm.com/T34550