debian: Build-Depend on systemd
systemd.pc is used to look up the paths to install .service files and
tmpfiles.d snippets.

(Now, you might ask yourself: why do we need to look these up when
systemd requires /lib to be a symlink to /usr/lib and does not allow
changing these paths? But that's one for another day.)

https://phabricator.endlessm.com/T35174