Log when powering off due to SIGTERM/SIGINT
I found it surprising that midway through a system shutdown, there was a
message in the journal indicating that payg_system_poweroff() had been
called. I wondered if this was a logging race, with eos-paygd
responsible for the shutdown occurring. In fact this is expected because
the daemon gets terminated during shutdown, and its shutdown request in
that case is redundant and ignored. (It is present in case the daemon is
terminated some other way.)

In most other cases where payg_system_poweroff is scheduled, a message
is logged explaining why. Log a message in this case too.

The one remaining case where payg_system_poweroff() is scheduled without
a log message is the 20-minute timeout for connecting to the system bus.
There is a g_debug() message there, but this is not enabled by default.
However this case is easier to distinguish.

https://phabricator.endlessm.com/T35468