efi: Simplify closing fds
libglnx provides an attribute you can attach to a variable holding a
file descriptor, which calls close() on it when it goes out of scope if
it is non-negative. Any error from close() is ignored, just like our
existing code, except if it is EBADF, which indicates a programming
error.

Use glnx_autofd throughout our EFI code. Most cases are simple but
eospayg_efi_init() is a little bit more involved.

This will make it easier to propagate errors to the caller in a future
commit.

https://phabricator.endlessm.com/T35459