efi: Fix eospayg_efi_PK_size() docstring
The implementation of efivarfs_read() returns in its size parameter the
length of the variable data, excluding the 4 bytes at the start of the
virtual file that represent its attributes. Failure to read those 4
bytes causes the size parameter to be set to -1.

eospayg_efi_PK_size() returns that size excluding the 4-byte attributes,
and its two call-sites either assume that behaviour or do not care:

    if (eospayg_efi_PK_size () != -1)
    if (eospayg_efi_PK_size () < 1)

but previously it was documented to return the size including EFI
overhead, which I infer means 4 bytes of attributes. Correct this to
match the implementation & expectation of its callers.

https://phabricator.endlessm.com/T35459