efi: Check size in eospayg_efi_var_read()
Previously, every caller of eospayg_efi_var_read(), except two, would
immediately check whether the returned size is equal to an expected
size, and discard the result if not.

Move this check into eospayg_efi_var_read(), with a new expected_size
parameter indicating the expected size of the returned data (or -1 to
skip this check). If this parameter is non-negative and the variable
contents do not match the expected size, NULL is returned rather than
the variable data (and, in future, an error code will be returned in an
additional parameter).

https://phabricator.endlessm.com/T35459