Revert "Partially revert "tests: Tell Meson our tests emit TAP output""
This reverts commit 2f1171b16ecd5e9d4e042e8faac1eae452a6340c, which
reverted part of commit 15f0d9319f29c56cdd345b67892013786084921b.

We now have a newer GLib which emits the "TAP version XX" header, so
let's try again to tell Meson that the C tests emit TAP output, and bump
the GLib dependency to the version that added that header.