efi: Check init() has been called in all public functions
Previously, many eospayg_efi_*() functions would assume that the `efi`
vtable is not NULL, and crash if this is the case. `efi` is initialised
by eospayg_efi_init(). It actually does this even if the function fails
which I think is somewhat weird, but, OK.

Add a

    g_return_[val_]if_fail (efi != NULL[, FALSE]);

guard to every function exposed in `efi.h` to catch this programmer
error.

https://phabricator.endlessm.com/T34898