efi: Guard against calling init() twice
Previously, eospayg_efi_init() would return early if an `inited` static
variable is TRUE. However this variable was never set.

Instead, check whether the `efi` static variable is non-NULL. `efi`
points to a vtable and is initialised on all calls to
eospayg_efi_init(), whether or not the function succeeds.

https://phabricator.endlessm.com/T34898